Update jaeger and opentracing versions.
Latest jaeger uses opentracing 0.30 which has incompatible changes
with previous versions.

+++ tchannel-core/src/test/java/com/uber/tchannel/tracing/TracingPropagationTest.java
@@ -335,7 +335,7 @@ public class TracingPropagationTest {
         String baggage = "Baggage-" + System.currentTimeMillis();
         span.setBaggageItem(BAGGAGE_KEY, baggage);
         if (!sampled) {
-            Tags.SAMPLING_PRIORITY.set(span, (short) 0);
+            Tags.SAMPLING_PRIORITY.set(span, 0);
         }
         tracingContext.pushSpan(span);
